requirements.txt
openai==1.12.0
google-auth==2.17.3
google-auth-oauthlib==0.7.1
google-auth-httplib2==0.1.0
google-api-python-client==2.116.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
beautifulsoup4==4.12.2
selenium==4.15.2
pydantic==2.5.3
loguru==0.7.2